refactor IDiagnostics and refine wording
